requirements.txt
astroid==2.0.4
Babel==2.6.0
colorama==0.3.9
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.1
django-pdfkit==0.3.1
django-phonenumber-field==2.0.1
django-wkhtmltopdf==3.2.0
gunicorn==19.9.0
html5lib==1.0.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pdfkit==0.6.0
phonenumbers==8.9.14
Pillow==5.3.0
psycopg2==2.7.5
pylint==2.1.1
PyPDF2==1.26.0
python-pdf==0.36
pytz==2018.5
reportlab==3.5.9
six==1.11.0
static3==0.7.0
webencodings==0.5.1
wrapt==1.10.11
xhtml2pdf==0.2.3
